JavaScript rot13 and rot47 enciphering scripts.
These were just hacked up quickly, originally intended to be executed
directly within a browser's address bar. They are certainly not written
with best practices in mind. I intentionally used very short variable
names and global variables for brevity. Also, I like the obscurity of just
a ton of letters and symbols all together. In any case, it should be
relatively easy to give variables function-scope and give the functions
more descriptive names. That is to say, if you actually want to use this
for something and know a little JavaScript it shouldn't be hard to make it
work.