Add a premake4 generated Makefile for convenience.
This file will need to be copied to the project root to be used, but it's
easier than going and installing premake4 if you're in a hurry (or just
don't want to). I am a little bit conflicted about my decision to use
premake4 since it still hasn't been packaged by Fedora. Maybe that is just
a sign that I should attempt to do so...